+++ src/roboto/query/__init__.py
@@ -5,8 +5,6 @@
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 from .api import (
-    MetadataValuesRequest,
-    MetadataValuesResponse,
     QualifiedRoboqlQuery,
     QueryContentMode,
     QueryContext,
@@ -48,8 +46,6 @@ __all__ = (
     "ConditionValue",
     "ConditionVisitor",
     "Field",
-    "MetadataValuesRequest",
-    "MetadataValuesResponse",
     "QualifiedRoboqlQuery",
     "Query",
     "QueryClient",